Update dependencies to reflect new Qiskit structure
The preferred install path for Qiskit is now the `qiskit` Python
package, and the "component" `qiskit-terra` is no longer preferred and
expected to be removed with Qiskit 1.0.  This updates the dependencies
to refer to the minimum supported version of the `qiskit` package
instead; the lowest version of that will now install more than is
completely required, but all latest versions are correct, and will be
correct moving forwards.