Fix layout output with hardware qubits
This fixes the output of programs that involve hardware qubits to the
form more standard to how Qiskit represents physical circuits.  Now, the
output circuit will have as many qubits as the maximum hardware qubit
explicitly used, even if the lower qubits are not used.  The
`TranspileLayout` will be a mapping of the qubits to their own indices,
to indicate that there were no underlying virtual qubits, but to assist
with other tools in recognising that the circuit is defined on physical
qubits.

The previous handling of hardware qubits created a circuit with only
explicitly referenced qubits in the circuit, then attempted to convey
the information about the actual hardware indices via the
`TranspileLayout`.  Unfortunately, this was not in a form that was
readily understandable to other tools, and the model of the "physical"
circuit not directly having its bit indices correspond to the hardware
qubit index was at odds with Qiskit's usual model of physical circuits.