Allow float/angle interaction in gate-parameter contexts (#35)
* Allow float/angle interaction in gate-parameter contexts

This allows technically invalid OpenQASM 3, but is in practice what
Qiskit's OpenQASM 3 exporter outputs due to problems in how it
represents symbolic expressions.  It's probably better for users to
allow the technically invalid OpenQASM 3 than to punish them for things
that are actually Qiskit's fault.

* Allow `float`/`angle` interaction in addition/subtraction

This is the same as the multiplication/division part. While it's much
less likely for these to appear from the Qiskit export, it makes sense
to allow them in non-strict mode for symmetry.

* Comment on strict mode